+++ lib.commonjs/abi/fragments.js
@@ -429,7 +429,7 @@ class ParamType {
             format = "sighash";
         }
         if (format === "json") {
-            const name = this.name || undefined; // @TODO: Make this "" (minor bump)
+            const name = this.name || "";
             if (this.isArray()) {
                 const result = JSON.parse(this.arrayChildren.format("json"));
                 result.name = name;