+++ lib.esm/contract/contract.js
@@ -3,7 +3,7 @@ import { isAddressable, resolveAddress } from "../address/index.js";
 // import from provider.ts instead of index.ts to prevent circular dep
 // from EtherscanProvider
 import { copyRequest, Log } from "../providers/provider.js";
-import { defineProperties, getBigInt, isCallException, isHexString, resolveProperties, makeError, assert, assertArgument } from "../utils/index.js";
+import { defineProperties, getBigInt, isCallException, isHexString, resolveProperties, isError, makeError, assert, assertArgument } from "../utils/index.js";
 import { ContractEventPayload, ContractUnknownEventPayload, ContractTransactionResponse, EventLog } from "./wrappers.js";
 const BN_0 = BigInt(0);
 function canCall(value) {
@@ -572,17 +572,20 @@ export class BaseContract {
         setInternal(this, { addrPromise, addr, deployTx, subs });
         // Add the event filters
         const filters = new Proxy({}, {
-            get: (target, _prop, receiver) => {
+            get: (target, prop, receiver) => {
                 // Pass important checks (like `then` for Promise) through
-                if (passProperties.indexOf(_prop) >= 0) {
-                    return Reflect.get(target, _prop, receiver);
+                if (typeof (prop) === "symbol" || passProperties.indexOf(prop) >= 0) {
+                    return Reflect.get(target, prop, receiver);
                 }
-                const prop = String(_prop);
-                const result = this.getEvent(prop);
-                if (result) {
-                    return result;
+                try {
+                    return this.getEvent(prop);
                 }
-                throw new Error(`unknown contract event: ${prop}`);
+                catch (error) {
+                    if (!isError(error, "INVALID_ARGUMENT") || error.argument !== "key") {
+                        throw error;
+                    }
+                }
+                return undefined;
             },
             has: (target, prop) => {
                 // Pass important checks (like `then` for Promise) through
@@ -598,22 +601,26 @@ export class BaseContract {
         });
         // Return a Proxy that will respond to functions
         return new Proxy(this, {
-            get: (target, _prop, receiver) => {
-                if (_prop in target || passProperties.indexOf(_prop) >= 0 || typeof (_prop) === "symbol") {
-                    return Reflect.get(target, _prop, receiver);
+            get: (target, prop, receiver) => {
+                if (typeof (prop) === "symbol" || prop in target || passProperties.indexOf(prop) >= 0) {
+                    return Reflect.get(target, prop, receiver);
                 }
-                const prop = String(_prop);
-                const result = target.getFunction(prop);
-                if (result) {
-                    return result;
+                // Undefined properties should return undefined
+                try {
+                    return target.getFunction(prop);
                 }
-                throw new Error(`unknown contract method: ${prop}`);
+                catch (error) {
+                    if (!isError(error, "INVALID_ARGUMENT") || error.argument !== "key") {
+                        throw error;
+                    }
+                }
+                return undefined;
             },
             has: (target, prop) => {
-                if (prop in target || passProperties.indexOf(prop) >= 0 || typeof (prop) === "symbol") {
+                if (typeof (prop) === "symbol" || prop in target || passProperties.indexOf(prop) >= 0) {
                     return Reflect.has(target, prop);
                 }
-                return target.interface.hasFunction(String(prop));
+                return target.interface.hasFunction(prop);
             }
         });
     }